refactor(mixin): 重构并移动多个mixin类以移除forge包依赖
- 将 FallingBlockEntityMixin 从 forge 包移出并更新包声明
- 移除 FallingBlockEntity Mixin 中的 @SuppressWarnings 注解
- 删除 LevelChunkMixin 的 forge 版本并将其功能合并到主类中
- 更新 LevelChunkMixin 以包含完整的方块实体加载和卸载事件处理
- 将 LightningBoltMixin 从 forge 包移出并更新包声明
- 在 anvilcraft.mixins.json 中更新 mixin 类引用路径
- 移除 forge 包下的 package-info.java 文件
- 添加缺失的导入和注解以支持新的 mixin 实现
