--- src/main/java/dev/dubhe/anvilcraft/mixin/forge/package-info.java
@@ -1,7 +1,0 @@
-@MethodsReturnNonnullByDefault
-@ParametersAreNonnullByDefault
-package dev.dubhe.anvilcraft.mixin.forge;
-
-import net.minecraft.MethodsReturnNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
