+++ src/main/resources/anvilcraft.mixins.json
@@ -26,6 +26,7 @@
     "EntityMixin",
     "EntityTypeBuilderMixin",
     "ExplosionMixin",
+    "FallingBlockEntityMixin",
     "FlyingHitEntityMixin",
     "HolderReferenceMixin",
     "HopperBlockMixin",
@@ -34,6 +35,7 @@
     "ItemEntityMixin",
     "ItemMixin",
     "ItemStackMixin",
+    "LightningBoltMixin",
     "LivingEntityMixin",
     "NearestAttackableTargetGoalMixin",
     "PhantomGoalMixin",
@@ -64,9 +66,6 @@
     "compat.AE2BatchCrafterBlockMixin",
     "compat.CerbonNewBeaconMenuMixin",
     "compat.CreateContraptionMixin",
-    "forge.FallingBlockEntityMixin",
-    "forge.LevelChunkMixin",
-    "forge.LightningBoltMixin",
     "invoker.BlockBehaviourInvoker",
     "invoker.BlockItemInvoker",
     "piglin.AbstractPiglinMixin",
